Refactor layout structure and update CSS files
The commit moves layout related tags from individual pages to a main layout component, thus greatly simplifying page structure. Also, the CSS file of the page has been removed, indicating a change in styling or removal of redundant styles. A new file for a "Projects" page was added as well.